% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_data.R
\name{cache_data}
\alias{cache_data}
\title{Cache an object into a file and load it into global environment}
\usage{
cache_data(
  fpath,
  object,
  fun_load = readRDS,
  fun_save = function(x, y) saveRDS(object = x, file = y)
)
}
\arguments{
\item{fpath}{file path}

\item{object}{an object or an expression which return object}

\item{fun_load}{a function to load object from file}

\item{fun_save}{a function to write object into file.}
}
\description{
Cache an object into a file and load it into global environment
}
\examples{
\dontrun{
# cache iris to the file
cache_data("iris.rds", iris)

# cache head or iris to the file
cache_data("iris_head.rds", {
  iris |>
    head(10)
})
iris_head

# cache head of iris to the file with save and load function
cache_data("iris_head.csv",
  fun_save = \(x, y) write.csv(x, y),
  fun_load = read.csv,
  {
    iris |>
      head(10)
  }
)
iris_head
}

}
