% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{cache_io}
\alias{cache_io}
\alias{cache_write}
\alias{cache_read}
\title{Read/Write cached command-line outputs.}
\usage{
cache_write(
  x,
  wd,
  ...,
  create_dir = TRUE,
  save_dir = getOption("ngsmisc.path_cmdout.save_dir", "cmdout_cache")
)

cache_read(
  wd,
  ...,
  save_dir = getOption("ngsmisc.path_cmdout.save_dir", "cmdout_cache")
)
}
\arguments{
\item{x}{A character vector to write to a cache file.}

\item{wd}{A path or character vector to the working directory.}

\item{...}{Arguments passed on to \code{fs::path()}.}

\item{create_dir}{A logical. if set to \code{TRUE}, will create directory of the
created file path. (default: \code{TRUE})}

\item{save_dir}{A path or character vector of directory name to save the
cache files. (default: \code{"cmdout_cache"}, default can be overridden by
setting \code{options("ngsmisc.path_cmdout.save_dir")}.)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Wrapper functions of \code{path_cmdout()} to help the text file I/O.
}
\examples{
NULL

}
