% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcount.R
\name{calc_tpm}
\alias{calc_tpm}
\title{Calculate Transcripts Per kilobase Million (TPM)}
\usage{
calc_tpm(readcount, len)
}
\arguments{
\item{readcount}{read count. atomic vector}

\item{len}{feature length. atomic vector}
}
\description{
Calculate Transcripts Per kilobase Million (TPM)
}
