% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{cat_stderr}
\alias{cat_stderr}
\title{Show stderr messages in the process output}
\usage{
cat_stderr(ps_out)
}
\arguments{
\item{ps_out}{output from processx::run()}
}
\description{
Show stderr messages in the process output
}
