% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{cat_stdout}
\alias{cat_stdout}
\title{Show stdout messages in the process output}
\usage{
cat_stdout(ps_out)
}
\arguments{
\item{ps_out}{output from processx::run()}
}
\description{
Show stdout messages in the process output
}
