% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{clP_is_significant}
\alias{clP_is_significant}
\title{Check the ego contains significantly enriched GO term}
\usage{
clP_is_significant(ego)
}
\arguments{
\item{ego}{An enrichResult class object.}
}
\description{
Check the ego contains significantly enriched GO term
}
\keyword{internal}
