% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{clP_num_significant}
\alias{clP_num_significant}
\title{Return the number of significantly enriched GO term in the ego}
\usage{
clP_num_significant(ego)
}
\arguments{
\item{ego}{An enrichResult class object.}
}
\description{
Return the number of significantly enriched GO term in the ego
}
\keyword{internal}
