% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{clP_path2label}
\alias{clP_path2label}
\title{Extract label of gene set from the file path of result list}
\usage{
clP_path2label(path_li_ego_rds)
}
\arguments{
\item{path_li_ego_rds}{Path to RDS files of the li_ego object.}
}
\description{
Extract label of gene set from the file path of result list
}
