% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{clP_wrap_barplots}
\alias{clP_wrap_barplots}
\title{Wrapping barplots of enrichResult-objects}
\usage{
clP_wrap_barplots(li_gp, scale_fill = ggplot2::scale_fill_viridis_c)
}
\arguments{
\item{li_gp}{list of ggplot objects}

\item{scale_fill}{A function controlling fill color of bars. it must be compatible with ggplot2::scale_fill_*(). default: ggplot2::scale_fill_viridis_c}
}
\description{
Wrapping barplots of enrichResult-objects
}
\examples{
li_gp <-
  system.file(package = "ngsmisc", "clusterprofiler", "li_ego_example.rds") |>
  readRDS() |>
  clP_plot_bar()

li_gp |> clP_wrap_barplots() &
  ggplot2::theme(text = ggplot2::element_text(size = 8))

li_gp |> clP_wrap_barplots(\(...) ggplot2:: scale_fill_gradientn(..., colours = terrain.colors(10))) &
  ggplot2::theme(text = ggplot2::element_text(size = 8))

}
