% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{clP_write_li_ego_as_bar}
\alias{clP_write_li_ego_as_bar}
\title{Write the enrichGO result to a barplot}
\usage{
clP_write_li_ego_as_bar(path_li_ego_rds)
}
\arguments{
\item{path_li_ego_rds}{Path to RDS files of the li_ego object.}
}
\description{
Write the enrichGO result to a barplot
}
