% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{clP_write_li_ego_as_csv}
\alias{clP_write_li_ego_as_csv}
\title{Write the enrichGO result to a csv file}
\usage{
clP_write_li_ego_as_csv(path_li_ego_rds)
}
\arguments{
\item{path_li_ego_rds}{Path to RDS files of the li_ego object.}
}
\description{
Write the enrichGO result to a csv file
}
\examples{
#inf <- system.file(package = "ngsmisc", "clusterprofiler", "li_ego_example.rds")
#clP_write_li_ego_as_csv(inf)

}
