% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{common_args_docker_helpler_fns}
\alias{common_args_docker_helpler_fns}
\title{common arguments for docker_helpler_fns.R}
\arguments{
\item{statement}{a command line statement.}

\item{sep_cmd}{separated command created by \code{tokenize_cmd()}.}

\item{ps_out}{an output from \code{processx::run()}}

\item{...}{further options are passed to \code{processx::run()}.}
}
\description{
common arguments for docker_helpler_fns.R
}
\keyword{internal}
