% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_ddr_plot_independent_filtering}
\alias{ds2_ddr_plot_independent_filtering}
\title{Plot independent filtering}
\usage{
ds2_ddr_plot_independent_filtering(ddr, title = "")
}
\arguments{
\item{ddr}{a DESeqResult object}

\item{title}{plot title}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{ds2_ddr_plot_independent_filtering()} plots a independent filtering result.
}
