% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_ddr_to_tbl}
\alias{ds2_ddr_to_tbl}
\title{Convert a DESeqResults object to tibble}
\usage{
ds2_ddr_to_tbl(ddr, rownames = "Geneid")
}
\arguments{
\item{ddr}{a \code{DESeq2::DESeqResults} class object object}

\item{rownames}{a column name. (default: \code{"Geneid"})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{ds2_ddr_to_tbl()} converts from a \code{DESeq2::DESeqResults} class object to a tibble.
}
