% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_estimate_disp}
\alias{ds2_dds_estimate_disp}
\title{Estimate dispersions}
\usage{
ds2_dds_estimate_disp(dds, ...)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{...}{further arguments are passed to \code{DESeq2::estimateDispersions()}. see Details.}
}
\description{
\code{ds2_dds_estimate_disp()} does the same thing as \code{DESeq2::estimateDispersions()}.
}
\details{
To see the advanced settings, run \code{?DESeq2::`estimateDispersions,DESeqDataSet-method`}.
}
\examples{
if(requireNamespace("DESeq2", quietly = TRUE)) {
  # Example read-count data.frame
  rcdf <-
    system.file(package = "ngsmisc", "deseq2") |>
    fs::dir_ls(regexp = "count.csv$") |>
    readr::read_csv(show_col_types = FALSE) |>
    ds2_tbl_to_rcdf()

  # Prepare `DESeqDataSet` object, estimate size-factor, and compute dispersion.
  dds <-
    ds2_rcdf_to_dds(
      rcdf = rcdf,
      coldata = data.frame(sample = colnames(rcdf))
    ) |>
    ds2_dds_estimate_sizefactor() |>
    ds2_dds_estimate_disp()

  dds
}

}
