% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_estimate_sizefactor}
\alias{ds2_dds_estimate_sizefactor}
\title{Estimate library size factors}
\usage{
ds2_dds_estimate_sizefactor(dds, ...)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{...}{...}
}
\description{
Estimate library size factors
}
