% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_get_normalized_count_tbl}
\alias{ds2_dds_get_normalized_count_tbl}
\title{Get normalized count data as a tibble}
\usage{
ds2_dds_get_normalized_count_tbl(dds, rownames = "Geneid")
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{rownames}{a column name. default: "Geneid"}
}
\description{
Get normalized count data as a tibble
}
