% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_get_sizefactor}
\alias{ds2_dds_get_sizefactor}
\title{Get library sizeFactors}
\usage{
ds2_dds_get_sizefactor(dds, ...)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{...}{...}
}
\description{
Get library sizeFactors
}
