% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_set_design}
\alias{ds2_dds_set_design}
\title{Set experimental design}
\usage{
ds2_dds_set_design(dds, design)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{design}{design}
}
\description{
Set experimental design
}
