% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_set_sizefactor}
\alias{ds2_dds_set_sizefactor}
\title{Set library size factors manually}
\usage{
ds2_dds_set_sizefactor(dds, sizefactor)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{sizefactor}{a double vector. its length must be equal to ncol(dds)}
}
\description{
Set library size factors manually
}
