% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_test_nbinomLRT}
\alias{ds2_dds_test_nbinomLRT}
\title{Run Likelihood ratio test for GLMs}
\usage{
ds2_dds_test_nbinomLRT(dds, reduced = ~1)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{reduced}{reduced model}
}
\description{
Run Likelihood ratio test for GLMs
}
