% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_test_nbinomWaldTest}
\alias{ds2_dds_test_nbinomWaldTest}
\title{Run Wald test for GLM coefficients}
\usage{
ds2_dds_test_nbinomWaldTest(dds, ...)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{...}{...}
}
\description{
Run Wald test for GLM coefficients
}
