% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_to}
\alias{ds2_dds_to}
\alias{ds2_dds_to_ddr}
\alias{ds2_dds_to_tbl}
\title{Convert \code{DESeq2::DESeqDataSet} class object to other object}
\usage{
ds2_dds_to_ddr(dds, ...)

ds2_dds_to_tbl(dds, rownames = "Geneid")
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{...}{...}

\item{rownames}{a column name. (default: \code{"Geneid"})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{ds2_dds_to_ddr()} converts to a \code{DESeq2::DESeqResults} class object.

\code{ds2_dds_to_tbl()} converts to a tibble.
}
