% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_to_ddr}
\alias{ds2_dds_to_ddr}
\title{Get a DESeqResults object from a DESeqDataSet object}
\usage{
ds2_dds_to_ddr(dds, ...)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{...}{...}
}
\description{
Get a DESeqResults object from a DESeqDataSet object
}
