% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_to_tbl}
\alias{ds2_dds_to_tbl}
\title{Convert a DESeqDataSet object to tibble}
\usage{
ds2_dds_to_tbl(dds, rownames = "Geneid")
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{rownames}{a column name. default: "Geneid"}
}
\description{
Convert a DESeqDataSet object to tibble
}
