% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_design}
\alias{ds2_design}
\alias{ds2_dds_get_design}
\alias{ds2_dds_set_design}
\title{Get/Set the experimental design}
\usage{
ds2_dds_get_design(dds)

ds2_dds_set_design(dds, design)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{design}{design}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{ds2_dds_get_design()} gets design.

\code{ds2_dds_set_design()} sets design.
}
