% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_mat_check_linear_dependency}
\alias{ds2_mat_check_linear_dependency}
\title{Find linear dependence in between columns in the matrix}
\usage{
ds2_mat_check_linear_dependency(mat, to_colname = TRUE)
}
\arguments{
\item{mat}{A matrix.}

\item{to_colname}{If \code{TRUE}, the colnames are used as variable names. (default: \code{TRUE})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{ds2_mat_check_linear_dependency()} checks linear dependencies in between columns in a matrix.
}
\examples{
df <-
  data.frame(
    genotype = c("WT", "mutant") \%>\% rep(each = 4),
    treatment = c(rep("mock", 2), rep("poison", 2)) \%>\% rep(2),
    time = c("0h", "2h") \%>\% rep(4)
  )
row.names(df) <- apply(df, 1, paste, collapse = "_")
df

mat_independent <- model.matrix(~ genotype * treatment * time, df)
ds2_mat_check_linear_dependency(mat_independent)

mat_dependent <- mat_independent[3:8,]
mat_dependent <- model.matrix(~ genotype * treatment * time, df[3:8,])
ds2_mat_check_linear_dependency(mat_dependent)

ds2_mat_check_linear_dependency(matrix(rep(1, 4), ncol = 2), FALSE)

}
