% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_rcdf_filter_by_rownames}
\alias{ds2_rcdf_filter_by_rownames}
\title{Filter readcount data.frame by rownames}
\usage{
ds2_rcdf_filter_by_rownames(rcdf, pattern, negate = TRUE)
}
\arguments{
\item{rcdf}{a data.frame created by ds2_prepare_rcdf()}

\item{pattern}{a regex pattern to look for.}

\item{negate}{If TRUE, return non-matching elements. default: TRUE}
}
\description{
Filter readcount data.frame by rownames
}
