% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_rcdf_filter_organella}
\alias{ds2_rcdf_filter_organella}
\title{Filter organella reads from readcount data.frame}
\usage{
ds2_rcdf_filter_organella(rcdf)
}
\arguments{
\item{rcdf}{a data.frame created by ds2_prepare_rcdf()}
}
\description{
Filter organella reads from readcount data.frame
}
