% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_rcdf_to_dds}
\alias{ds2_rcdf_to_dds}
\title{Prepare DESeqDataSet object from read-count data.frame}
\usage{
ds2_rcdf_to_dds(rcdf, coldata, design = ~1)
}
\arguments{
\item{rcdf}{a read-count data.frame created by \code{ds2_tbl_to_rcdf()}}

\item{coldata}{a data.frame specifying sample info}

\item{design}{design. (default: \code{~ 1} (no design))}
}
\description{
\code{ds2_rcdf_to_dds()} does the same thing as \code{DESeq2::DESeqDataSetFromMatrix()}.
}
\examples{
if(requireNamespace("DESeq2", quietly = TRUE)) {
  # Example read-count data.frame.
  rcdf <-
    system.file(package = "ngsmisc", "deseq2") |>
    fs::dir_ls(regexp = "count.csv$") |>
    readr::read_csv(show_col_types = FALSE) |>
    ds2_tbl_to_rcdf()

  # Convert read-count data.frame to a `DESeq2::DESeqDataSet-class` object.
  ds2_rcdf_to_dds(
    rcdf = rcdf,
    coldata = data.frame(sample = colnames(rcdf))
  )
}
}
