% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_sizefactor}
\alias{ds2_sizefactor}
\alias{ds2_dds_estimate_sizefactor}
\alias{ds2_dds_get_sizefactor}
\alias{ds2_dds_set_sizefactor}
\title{Manipulate (estimate/get/set) library size factors}
\usage{
ds2_dds_estimate_sizefactor(dds, ...)

ds2_dds_get_sizefactor(dds, ...)

ds2_dds_set_sizefactor(dds, sizefactor)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{...}{...}

\item{sizefactor}{a double vector. its length must be equal to ncol(dds)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{ds2_dds_estimate_sizefactor()} does the same thing as \code{DESeq2::estimateSizeFactors()}.

\code{ds2_dds_get_sizefactor()} gets size-factor as a named numeric vector.

\code{ds2_dds_set_sizefactor()} manually sets specified size-factor to the given object.
}
\examples{
# example read-count data.frame
rcdf <-
  system.file(package = "ngsmisc", "deseq2") \%>\%
  fs::dir_ls(regexp = "count.csv$") \%>\%
  {suppressMessages(readr::read_csv(.))} \%>\%
  ds2_tbl_to_rcdf()

dds_wo_sf <-
  ds2_rcdf_to_dds(
    rcdf = rcdf,
    coldata = data.frame(sample = colnames(rcdf))
  )

# estimate and set library size factors
dds_w_sf <- dds_wo_sf \%>\% ds2_dds_estimate_sizefactor()

# if the specfied `dds` object has not been calculate size-factor previously,
# automatically calculate size-factor and return it.
dds_wo_sf \%>\% ds2_dds_get_sizefactor()
dds_w_sf \%>\% ds2_dds_get_sizefactor()

dds_wo_sf \%>\% ds2_dds_set_sizefactor(1:3) \%>\% ds2_dds_get_sizefactor()

}
