% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_tbl_to_rcdf}
\alias{ds2_tbl_to_rcdf}
\title{Prepare input readcount data.frame for DESeq2 from a tibble}
\usage{
ds2_tbl_to_rcdf(tbl, genename_col = "Geneid")
}
\arguments{
\item{tbl}{a tibble containing read count columns and a gene name column}

\item{genename_col}{character specifying the gene name column. default: "Geneid"}
}
\description{
Prepare input readcount data.frame for DESeq2 from a tibble
}
