% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_test}
\alias{ds2_test}
\alias{ds2_dds_test_nbinomLRT}
\alias{ds2_dds_test_nbinomWaldTest}
\title{Run Likelihood ratio test or Wald test}
\usage{
ds2_dds_test_nbinomLRT(dds, reduced = ~1)

ds2_dds_test_nbinomWaldTest(dds, ...)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{reduced}{reduced model. (default: \code{~ 1})}

\item{...}{pass to the DESeq2::nbinomWaldTest()}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{ds2_dds_test_nbinomLRT()} does the same thing as \code{DESeq2::nbinomLRT()}.

\code{ds2_dds_test_nbinomWaldTest()} does the same thing as \code{DESeq2::nbinomWaldTest()}.
}
