% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_calc}
\alias{fC_calc}
\alias{fC_calc_rpm}
\alias{fC_calc_rpkm}
\alias{fC_calc_tpm}
\title{Calculate RPM/RPKM/TPM from featureCounts read counts}
\usage{
fC_calc_rpm(tbl_fC)

fC_calc_rpkm(tbl_fC)

fC_calc_tpm(tbl_fC)
}
\arguments{
\item{tbl_fC}{featureCounts read counts tibble}
}
\description{
Calculate RPM/RPKM/TPM from featureCounts read counts
}
\examples{
infs <-
  system.file(package = "ngsmisc") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt$")

lapply(infs, fC_read_count) \%>\%
  fC_merge_tbl_li() \%>\%
  fC_rename_col(col_fpath = -c(1:6)) \%>\%
  fC_calc_rpm() \%>\%
  fC_calc_rpkm() \%>\%
  fC_calc_tpm() \%>\%
  dplyr::glimpse()
}
