% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_count}
\alias{fC_count}
\alias{fC_read_count}
\alias{fC_merge_count}
\title{Read featureCounts read count table as tibbles and merge them}
\usage{
fC_read_count(fpath)

fC_merge_count(li_tbl)
}
\arguments{
\item{fpath}{featureCounts output file path}

\item{li_tbl}{A list of featureCounts read counts tibbles}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{fC_read_count()} reads a featureCounts read-count output.

\code{fC_merge_count()} merges a list of tibbles read by \code{fC_read_count()} to a tibble.
}
\examples{
# example read-count files
infs <-
  system.file(package = "ngsmisc", "featurecounts") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt$")

# show first six lines
readLines(infs[1], n = 6) \%>\% cat(sep = "\n")

# read a read-count file
fC_read_count(infs[1])

# read read-count files and merge them into a tibble
lapply(infs, fC_read_count) \%>\%
  lapply(rename_fpath_bam, nth = 7) \%>\%
  fC_merge_count()

}
