% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_read_count}
\alias{fC_read_count}
\title{Read featureCounts read count table}
\usage{
fC_read_count(fpath)
}
\arguments{
\item{fpath}{featureCounts output file path}
}
\description{
Read featureCounts read count table
}
\examples{
infs <-
  system.file(package = "ngsmisc") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt$")
fC_read_count(infs[1])

}
