% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_read_summary}
\alias{fC_read_summary}
\title{Read featureCounts read count summary}
\usage{
fC_read_summary(fpath)
}
\arguments{
\item{fpath}{featureCounts output file path}
}
\description{
Read featureCounts read count summary
}
\examples{
infs <-
  system.file(package = "ngsmisc") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt.summary$")
fC_read_summary(infs[1])

}
