% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_rename_col}
\alias{fC_rename_col}
\title{Rename featureCounts read count table}
\usage{
fC_rename_col(tbl_fC, col_fpath = 7L, file_suffix = ".sort.bam$")
}
\arguments{
\item{tbl_fC}{featureCounts read counts tibble}

\item{col_fpath}{integer. default = 7L}

\item{file_suffix}{regex pattern. default = ".sort.bam$"}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
infs <-
  system.file(package = "ngsmisc") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt$")
fC_read_count(infs[1]) \%>\% fC_rename_col()

}
