% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_summary}
\alias{fC_summary}
\alias{fC_read_summary}
\alias{fC_merge_summary}
\title{Read featureCounts read count summary as tibbles and merge them}
\usage{
fC_read_summary(fpath)

fC_merge_summary(li_tbl)
}
\arguments{
\item{fpath}{featureCounts summary output file path}

\item{li_tbl}{A list of featureCounts summary tibbles}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{fC_read_summary()} reads a featureCounts summary file.
}
\examples{
# example summary files
infs <-
  system.file(package = "ngsmisc", "featurecounts") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt.summary$")

# show first six lines
readLines(infs[1], n = 6) \%>\% cat(sep = "\n")

# read a summary file
fC_read_summary(infs[1])

# read summary files and merge them into a tibble
lapply(infs, fC_read_summary) \%>\%
  lapply(rename_fpath_bam, nth = 2) \%>\%
  fC_merge_summary()

}
