% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_write_count}
\alias{fC_write_count}
\title{Write selected columns (id/column/rpm/rpkm/tpm) in a tibble to a csv file.}
\usage{
fC_write_count(tbl_fC, fpath)
}
\arguments{
\item{tbl_fC}{a tibble read from featureCounts read-count output.}

\item{fpath}{an output csv file path.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{fC_write_count()} writes several columns in featureCounts read-count tibble to csv file.
}
\examples{
# example read-count files
infs <-
  system.file(package = "ngsmisc", "featurecounts") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt$")

# read read-count files and merge them into a tibble
lapply(infs, fC_read_count) \%>\%
  lapply(rename_fpath_bam, nth = 7) \%>\%
  fC_merge_count() \%>\%
  fC_write_count("./count.csv")
readr::read_csv("./count.csv")
fs::file_delete("./count.csv")

}
