% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Return a function to determin pretty integer break
https://www.r-bloggers.com/2019/11/setting-axes-to-integer-values-in-ggplot2/}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{A integer. Number of breaks.}

\item{...}{Supplementary arguments passed to base::pretty()}
}
\description{
Return a function to determin pretty integer break
https://www.r-bloggers.com/2019/11/setting-axes-to-integer-values-in-ggplot2/
}
