% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsmisc-package.R
\docType{package}
\name{ngsmisc-package}
\alias{ngsmisc}
\alias{ngsmisc-package}
\title{ngsmisc: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Toshihiro Arae \email{toshihiro.arae@gmail.com}

}
\keyword{internal}
