% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{path_cmdout}
\alias{path_cmdout}
\title{Create file path to save the command line output}
\usage{
path_cmdout(
  wd,
  ...,
  create_dir = TRUE,
  save_dir = getOption("ngsmisc.path_cmdout.save_dir", "cmdout_cache")
)
}
\arguments{
\item{wd}{A path or character vector to the working directory.}

\item{...}{Arguments passed on to \code{fs::path()}.}

\item{create_dir}{A logical. if set to \code{TRUE}, will create directory of the
created file path. (default: \code{TRUE})}

\item{save_dir}{A path or character vector of directory name to save the
cache files. (default: \code{"cmdout_cache"}, default can be overridden by
setting \code{options("ngsmisc.path_cmdout.save_dir")}.)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Make a file path by inserting a directory between the working directory and
specified file path.
}
\examples{
path_cmdout("/path/to/wd", "cmd_out.txt", create_dir = FALSE)
path_cmdout("/path/to/wd", "cmd_out.txt", create_dir = FALSE, save_dir = "other")
path_cmdout("/path/to/wd", "level1", "level2", "cmd_out.txt", create_dir = FALSE)

}
