% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_col_AGI}
\alias{rename_col_AGI}
\title{Rename tibble column names which containing AGI}
\usage{
rename_col_AGI(tbl, col_AGI)
}
\arguments{
\item{tbl}{data.frame.}

\item{col_AGI}{integer vector. specified column number which containing file path in its column name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
