% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_col_fpath}
\alias{rename_col_fpath}
\title{Rename tibble column names which containing file path}
\usage{
rename_col_fpath(tbl, col_fpath, rm_file_suffix, file_suffix)
}
\arguments{
\item{tbl}{data.frame.}

\item{col_fpath}{integer vector. specified column number which containing file path in its column name}

\item{rm_file_suffix}{logical. if TRUE, specified file suffix are removed from column name}

\item{file_suffix}{regex pattern which specified file suffix to remove}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
