% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_fpath}
\alias{rename_fpath}
\alias{rename_fpath_fastq}
\alias{rename_fpath_bam}
\title{Rename file path in the column name to the label}
\usage{
rename_fpath(tbl, nth, prefix = "", suffix = "")

rename_fpath_fastq(tbl, nth, prefix = "")

rename_fpath_bam(tbl, nth, prefix = "")
}
\arguments{
\item{tbl}{a data.frame to rename.}

\item{nth}{a numeric vector to specify the column index(es) to rename.}

\item{prefix}{a string to specify the removing file prefix. (default: \code{""})}

\item{suffix}{a string to specify the removing file suffix. (defaults are bellow)
\itemize{
\item for \code{rename_fpath()}: \code{""}
\item for \code{rename_fpath_fastq()}: \code{"([.](fastq|fq))([.]gz){0,1}$"}
\item for \code{rename_fpath_bam()}: \code{"([.]sort){0,1}([.]bam)$"}
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{rename_fpath_fastq()}, \code{rename_fpath_bam} are the renaming functions for
a column containing fastq and bam file paths, respectively.
}
\examples{
f <- function(...) data.frame(..., check.names = FALSE)

f("dir_log/sample1.log" = 1) \%>\% rename_fpath(nth = 1, suffix = ".log")

f("dir_fastq/sample1.fastq" = 1) \%>\% rename_fpath_fastq(nth = 1)
f("dir_fastq/sample1.fq" = 1) \%>\% rename_fpath_fastq(nth = 1)
f("dir_fastq/sample1.fastq.gz" = 1) \%>\% rename_fpath_fastq(nth = 1)
f("dir_fastq/sample1.fq.gz" = 1) \%>\% rename_fpath_fastq(nth = 1)

f("dir_bam/sample1.bam" = 1) \%>\% rename_fpath_bam(nth = 1)
f("dir_bam/sample1.sort.bam" = 1) \%>\% rename_fpath_bam(nth = 1)

}
