% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_nth_col}
\alias{rename_nth_col}
\title{Rename the Nth column name of a data.frame.}
\usage{
rename_nth_col(tbl, nth, renamer)
}
\arguments{
\item{tbl}{a data.frame to rename.}

\item{nth}{a numeric vector to specify the column index(es) to rename.}

\item{renamer}{a function to convert column name.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{rename_nth_col()} renames the Nth column name by applying a renamer function.
}
\examples{
infs <-
  system.file(package = "ngsmisc", "featurecounts") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt$")

tbl <- fC_read_count(infs[1]) \%>\% dplyr::select(1, 7)
tbl
rename_nth_col(tbl, 2, basename)
rename_nth_col(tbl, 2, function(x) substr(basename(x), 1, 8))

tbl_merge <-
  lapply(infs, fC_read_count) \%>\%
  fC_merge_count() \%>\%
  dplyr::select(!c(2:6))
tbl_merge
rename_nth_col(tbl_merge, 2:4, function(x) substr(basename(x), 1, 8))
rename_nth_col(tbl_merge, -c(1:2), function(x) substr(basename(x), 1, 8))

}
