% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{renamer_fpath}
\alias{renamer_fpath}
\title{Return a renaming function for extract labels from file name.}
\usage{
renamer_fpath(prefix = "", suffix = "")
}
\arguments{
\item{prefix}{a string to specify the removing file prefix. (default: \code{""})}

\item{suffix}{a string to specify the removing file suffix. (default: \code{""})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{renamer_fpath()} returns a utility function to extract a label from a file path.
}
\examples{
renamer <- renamer_fpath()
renamer("/path/to/the/example_file1.txt")

renamer <- renamer_fpath(prefix = "example_")
renamer("/path/to/the/example_file1.txt")

renamer <- renamer_fpath(prefix = "example_", suffix = ".txt")
renamer("/path/to/the/example_file1.txt")

}
