% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{rM_extract_event_type}
\alias{rM_extract_event_type}
\title{Extract event type from file path}
\usage{
rM_extract_event_type(path)
}
\arguments{
\item{path}{file path}
}
\description{
Extract event type from file path
}
