% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{rM_filter_by_AGI}
\alias{rM_filter_by_AGI}
\title{Filter MATS tibble by AGI}
\usage{
rM_filter_by_AGI(tbl_MATS, AGI)
}
\arguments{
\item{tbl_MATS}{A MATS event tibble}

\item{AGI}{Character vector containing AGI interested}
}
\description{
Filter MATS tibble by AGI
}
