% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{rM_make_tbl_splice_site}
\alias{rM_make_tbl_splice_site}
\title{Create tibble containing splice site info from tbl_MATS}
\usage{
rM_make_tbl_splice_site(tbl_MATS, event_type)
}
\arguments{
\item{tbl_MATS}{A MATS event tibble}

\item{event_type}{AS event type}
}
\description{
Create tibble containing splice site info from tbl_MATS
}
