% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{rM_mutate_count_stat}
\alias{rM_mutate_count_stat}
\title{Mutate a count stat column to MATS tibble}
\usage{
rM_mutate_count_stat(tbl_MATS, convert_func, stat_name)
}
\arguments{
\item{tbl_MATS}{A MATS event tibble}

\item{convert_func}{A function which receive a vector of count data and return a vector.}

\item{stat_name}{A string. if it is specified, it will be used for column name.}
}
\description{
Mutate a count stat column to MATS tibble
}
