% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{rM_read_MATS_test_out}
\alias{rM_read_MATS_test_out}
\title{Read rMATS output file (.MATS.JC.txt/.MATS.JCEC.txt)}
\usage{
rM_read_MATS_test_out(fpath)
}
\arguments{
\item{fpath}{path to file}
}
\description{
Read rMATS output file (.MATS.JC.txt/.MATS.JCEC.txt)
}
