% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS.R
\name{rM_select_switch_start_end_col}
\alias{rM_select_switch_start_end_col}
\title{Select some columns to plotting}
\usage{
rM_select_switch_start_end_col(tbl_MATS, event_type)
}
\arguments{
\item{tbl_MATS}{A MATS event tibble}

\item{event_type}{AS event type}
}
\description{
Select some columns to plotting
}
