% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{run_cat_stdout}
\alias{run_cat_stdout}
\title{Run a statement with showing stdout messages}
\usage{
run_cat_stdout(statement, ...)
}
\arguments{
\item{statement}{a command line statement. (character)}

\item{...}{options for pass to processx::run()}
}
\description{
Run a statement with showing stdout messages
}
