% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{run_cmd}
\alias{run_cmd}
\title{Run a separated command in external process}
\usage{
run_cmd(
  sep_cmd,
  ...,
  error_on_status = TRUE,
  echo = rlang::is_interactive(),
  spinner = rlang::is_interactive()
)
}
\arguments{
\item{sep_cmd}{separated command created by \code{tokenize_cmd()}.}

\item{...}{further options are passed to \code{processx::run()}.}

\item{error_on_status}{see \code{processx::run()}}

\item{echo}{see \code{processx::run()}}

\item{spinner}{see \code{processx::run()}}
}
\description{
Run a separated command in external process
}
\examples{
rlang::with_interactive({
  cmd_ok <- "echo 'hoge hoge'"
  run_cmd(tokenize_cmd(cmd_ok))
})

rlang::with_interactive({
  cmd_err <- "cat sonzai_shinai_file.txt"
  run_cmd(tokenize_cmd(cmd_err), error_on_status = FALSE)
})

str(run_cmd(tokenize_cmd(cmd_ok)))

str(run_cmd(tokenize_cmd(cmd_err), error_on_status = FALSE))

}
