% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{sep_by_blank}
\alias{sep_by_blank}
\title{Separate a statement by white spaces}
\usage{
sep_by_blank(statement)
}
\arguments{
\item{statement}{a command line statement. (character)}
}
\description{
Separate a statement by white spaces
}
\examples{
cmd <- "echo 'hoge hoge'"
sep_by_blank(cmd)

}
