% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{ST_final_log}
\alias{ST_final_log}
\alias{ST_read_final_log}
\alias{ST_merge_final_log}
\title{Read STAR final.log files as tibbles and merge them}
\usage{
ST_read_final_log(fpath)

ST_merge_final_log(li_tbl)
}
\arguments{
\item{fpath}{a path to the STAR final.log file}

\item{li_tbl}{a list of data.frame}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
infs <-
  system.file(package = "ngsmisc", "star") \%>\%
  fs::dir_ls(regexp = ".final.log$")
ST_read_final_log(infs[1]) \%>\% rename_fpath(nth = 3) \%>\% print(n = Inf)

infs \%>\%
  lapply(ST_read_final_log) \%>\%
  lapply(rename_fpath, nth = 3, suffix = ".final.log") \%>\%
  ST_merge_final_log()

}
