% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{tokenize_cmd}
\alias{tokenize_cmd}
\title{Separate a shell command statement by white spaces}
\usage{
tokenize_cmd(statement)
}
\arguments{
\item{statement}{a command line statement.}
}
\description{
Separate a shell command statement by white spaces
}
\examples{
tokenize_cmd("ls -l -a")
tokenize_cmd("echo 'hoge hoge'")

}
