% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{wrap_strings}
\alias{wrap_strings}
\title{Wrap strings to defined width.}
\usage{
wrap_strings(string, width = 40)
}
\arguments{
\item{string}{A vector of strings.}

\item{width}{A number of string width to wrap. (default: 40)}
}
\description{
Wrap strings to defined width.
}
