% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{cache_io}
\alias{cache_io}
\alias{cache_write}
\alias{cache_read}
\title{Read/Write cached command-line outputs.}
\usage{
cache_write(
  x,
  wd,
  ...,
  create_dir = TRUE,
  save_dir = getOption("ngsmisc.path_cmdout.save_dir", "cmdout_cache")
)

cache_read(
  wd,
  ...,
  create_dir = TRUE,
  save_dir = getOption("ngsmisc.path_cmdout.save_dir", "cmdout_cache")
)
}
\arguments{
\item{x}{A character vector to write to a cache file.}

\item{wd}{a path to the working directory}

\item{...}{...}

\item{create_dir}{logical. if \code{TRUE} create directory of the returned path. (default: \code{TRUE})}

\item{save_dir}{directory name to save the output. (default: \code{"cmdout_cache"})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
NULL

}
