% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{cat_process_out}
\alias{cat_process_out}
\alias{cat_stdout}
\alias{cat_stderr}
\title{Show the stdout/stderr message in processx::run() ouput.}
\usage{
cat_stdout(ps_out)

cat_stderr(ps_out)
}
\arguments{
\item{ps_out}{an output from \code{processx::run()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
