% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{clP_wrap_barplots}
\alias{clP_wrap_barplots}
\title{Wrapping barplots of enrichResult-objects}
\usage{
clP_wrap_barplots(li_gp, fill_fun = ggplot2::scale_fill_viridis_c)
}
\arguments{
\item{li_gp}{list of ggplot objects}

\item{fill_fun}{A function to colorize bar}
}
\description{
Wrapping barplots of enrichResult-objects
}
