% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{clP_write_li_ego}
\alias{clP_write_li_ego}
\title{Run clustreProfiler::enrichGO() and write R object to a file}
\usage{
clP_write_li_ego(fg, bg, out_dir, label, ...)
}
\arguments{
\item{fg}{A vector of foreground gene ids.}

\item{bg}{A vector of background gene ids.}

\item{out_dir}{Path to the output directory.}

\item{label}{A label of the set of foreground gene}

\item{...}{Named options will be passed to clusterProfiler::enrichGO(). without "gene", "universe", "ont", "pool"}
}
\description{
Run clustreProfiler::enrichGO() and write R object to a file
}
