% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{cmd_run}
\alias{cmd_run}
\title{Run a separated command in external process}
\usage{
cmd_run(sep_cmd, ...)
}
\arguments{
\item{sep_cmd}{separated command created by \code{tokenize_cmd()}.}

\item{...}{further options are passed to \code{processx::run()}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
