% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{cmd_run}
\alias{cmd_run}
\title{Run a separated command in external process}
\usage{
cmd_run(sep_cmd, ...)
}
\arguments{
\item{sep_cmd}{separated command created by \code{sep_by_blank()}}

\item{...}{options for pass to \code{processx::run()}}
}
\description{
Run a separated command in external process
}
\examples{
cmd_ok <- "echo 'hoge hoge'"
cmd_run(sep_by_blank(cmd_ok))

cmd_err <- "cat sonzai_shinai_file.txt"
cmd_run(sep_by_blank(cmd_err), error_on_status = FALSE)

}
