% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_ddr_to_tbl}
\alias{ds2_ddr_to_tbl}
\title{Convert a DESeqResults object to a tibble}
\usage{
ds2_ddr_to_tbl(ddr, rownames = "Geneid")
}
\arguments{
\item{ddr}{a \code{DESeq2::DESeqResults} class object object}

\item{rownames}{a column name. (default: \code{"Geneid"})}
}
\description{
\code{ds2_ddr_to_tbl()} converts from a \code{DESeq2::DESeqResults} class object to a tibble.
}
\examples{
if(requireNamespace("DESeq2", quietly = TRUE)) {
  # Example read-count data.frame
  rcdf <-
    system.file(package = "ngsmisc", "deseq2") |>
    fs::dir_ls(regexp = "count.csv$") |>
    readr::read_csv(show_col_types = FALSE) |>
    ds2_tbl_to_rcdf()

  # Prepare `DESeqDataSet` object, estimate size-factor, and compute dispersion.
  dds <-
    ds2_rcdf_to_dds(
      rcdf = rcdf,
      coldata = data.frame(
        sample = colnames(rcdf),
        group = as.factor(c("A", "A", "B"))
      ),
      design = ~ group
    ) |>
    ds2_dds_estimate_sizefactor() |>
    ds2_dds_estimate_disp()

  # Likelihood ratio test
  ddr <- ds2_dds_test_nbinomLRT(dds) |> ds2_dds_to_ddr()

  # Convert `DESeqResults` to tibble
  ddr |> ds2_ddr_to_tbl()
}

}
