% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_ddr_to_tbl}
\alias{ds2_ddr_to_tbl}
\title{Convert a DESeqResults object to tibble}
\usage{
ds2_ddr_to_tbl(ddr, rownames = "Geneid")
}
\arguments{
\item{ddr}{a DESeqResult object}

\item{rownames}{a column name. default: "Geneid"}
}
\description{
Convert a DESeqResults object to tibble
}
