% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_dds_to}
\alias{ds2_dds_to}
\alias{ds2_dds_to_ddr}
\alias{ds2_dds_to_tbl}
\title{Convert \code{DESeq2::DESeqDataSet} class object to other class object}
\usage{
ds2_dds_to_ddr(dds, ...)

ds2_dds_to_tbl(dds, rownames = "Geneid")
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{...}{further arguments are passed to \code{DESeq2::results()}. See Details.}

\item{rownames}{a column name. (default: \code{"Geneid"})}
}
\description{
\code{ds2_dds_to_ddr()} converts to a \code{DESeq2::DESeqResults} class object.

\code{ds2_dds_to_tbl()} converts to a tibble.
}
\details{
To see the advanced settings, run \code{?DESeq2::results()}.
}
\examples{
if(requireNamespace("DESeq2", quietly = TRUE)) {
  # Example read-count data.frame
  rcdf <-
    system.file(package = "ngsmisc", "deseq2") |>
    fs::dir_ls(regexp = "count.csv$") |>
    readr::read_csv(show_col_types = FALSE) |>
    ds2_tbl_to_rcdf()

  # Prepare `DESeqDataSet` object, estimate size-factor, and compute dispersion.
  dds <-
    ds2_rcdf_to_dds(
      rcdf = rcdf,
      coldata = data.frame(
        sample = colnames(rcdf),
        group = as.factor(c("A", "A", "B"))
      ),
      design = ~ group
    ) |>
    ds2_dds_estimate_sizefactor() |>
    ds2_dds_estimate_disp()

  ## Convert `DESeqDataSet` to `DESeqResults` or tibble
  # Likelihood ratio test
  ds2_dds_test_nbinomLRT(dds) |> ds2_dds_to_ddr()
  ds2_dds_test_nbinomLRT(dds) |> ds2_dds_to_tbl()

  # Wald test
  ds2_dds_test_nbinomWaldTest(dds) |> ds2_dds_to_ddr()
  ds2_dds_test_nbinomWaldTest(dds) |> ds2_dds_to_tbl()
}

}
