% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_rcdf_filter_by_rownames}
\alias{ds2_rcdf_filter_by_rownames}
\title{Filter read-count data.frame by rownames}
\usage{
ds2_rcdf_filter_by_rownames(rcdf, pattern, negate = FALSE)
}
\arguments{
\item{rcdf}{a data.frame created by \code{ds2_tbl_to_rcdf()}}

\item{pattern}{a regex pattern to look for.}

\item{negate}{If TRUE, return non-matching elements. (default: \code{FALSE})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{ds2_rcdf_filter_by_rownames()} filter a read-count data.frame by rownames.
}
\examples{
# example read-count data.frame
rcdf <-
  system.file(package = "ngsmisc", "deseq2") \%>\%
  fs::dir_ls(regexp = "count.csv$") \%>\%
  {suppressMessages(readr::read_csv(.))} \%>\%
  ds2_tbl_to_rcdf()

head(rcdf)
ds2_rcdf_filter_by_rownames(head(rcdf), pattern = "[123]0")
ds2_rcdf_filter_by_rownames(head(rcdf), pattern = "[123]0", negate = TRUE)

}
