% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_rcdf_filter_organella}
\alias{ds2_rcdf_filter_organella}
\title{Filter organella reads from readcount data.frame}
\usage{
ds2_rcdf_filter_organella(rcdf)
}
\arguments{
\item{rcdf}{a read-count data.frame created by \code{ds2_tbl_to_rcdf()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
