% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{ds2_rcdf_to_dds}
\alias{ds2_rcdf_to_dds}
\title{Prepare DESeqDataSet object from readcount data.frame}
\usage{
ds2_rcdf_to_dds(rcdf, coldata, design = ~1)
}
\arguments{
\item{rcdf}{a data.frame created by ds2_prepare_rcdf()}

\item{coldata}{a data.frame specifying sample info}

\item{design}{design. default: ~ 1 (no design)}
}
\description{
Prepare DESeqDataSet object from readcount data.frame
}
