% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_merge_tbl_li}
\alias{fC_merge_tbl_li}
\title{Merge featureCounts read count table to a tibble}
\usage{
fC_merge_tbl_li(tbl_li_fC)
}
\arguments{
\item{tbl_li_fC}{list of featureCounts read counts tibbles}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
infs <-
  system.file(package = "ngsmisc") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt$")
lapply(infs, fC_read_count) \%>\% fC_merge_tbl_li()


}
