% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{fC_read_count}
\alias{fC_read_count}
\title{Read featureCounts read count table}
\usage{
fC_read_count(fpath)
}
\arguments{
\item{fpath}{featureCounts output file path}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
infs <-
  system.file(package = "ngsmisc", "featurecounts") \%>\%
  fs::dir_ls(regexp = "_gene_counts.txt$")
fC_read_count(infs[1])

}
