% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsmisc-package.R
\docType{package}
\name{ngsmisc-package}
\alias{ngsmisc}
\alias{ngsmisc-package}
\title{ngsmisc: Miscellaneous Functions for NGS Related Data Analysis}
\description{
Provides a collection of utility functions for processing and analyzing NGS related data, including quality control, data visualization, and statistical analysis tools.
}
\author{
\strong{Maintainer}: Toshihiro Arae \email{toshihiro.arae@gmail.com}

}
\keyword{internal}
