% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{path_cmdout}
\alias{path_cmdout}
\title{Create file path to save the command line output}
\usage{
path_cmdout(wd, ..., create_dir = TRUE, save_dir = "cmdout_cache")
}
\arguments{
\item{wd}{a path to the working directory}

\item{...}{...}

\item{create_dir}{logical. if TRUE create directory of the returned path. (default: TRUE)}

\item{save_dir}{directory name to save the output. (default: "cmdout_cache")}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
path_cmdout("/path/to/wd", "cmd_out.txt", create_dir = FALSE)
path_cmdout("/path/to/wd", "cmd_out.txt", create_dir = FALSE, save_dir = "other")
path_cmdout("/path/to/wd", "level1", "level2", "cmd_out.txt", create_dir = FALSE)

}
