% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_helper_fns.R
\name{run_cat_stderr}
\alias{run_cat_stderr}
\title{Run a command line and show stderr outputs from the result.}
\usage{
run_cat_stderr(statement, ...)
}
\arguments{
\item{statement}{a command line statement.}

\item{...}{further options are passed to \code{processx::run()}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
