% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{ST_final_log}
\alias{ST_final_log}
\alias{ST_read_final_log}
\alias{ST_merge_final_log}
\title{Read STAR final.log files and merge them into a tibble}
\usage{
ST_read_final_log(fpath, rename_col = function(x) x)

ST_merge_final_log(li_tbl)
}
\arguments{
\item{fpath}{a path to the STAR final.log file}

\item{rename_col}{a function to covert the column name of fpath. (default: function(x) x)}

\item{li_tbl}{a list of data.frame}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
infs <-
  system.file(package = "ngsmisc", "star") \%>\%
  fs::dir_ls(regexp = ".final.log$")
ST_read_final_log(infs[1], rename_col = fs::path_file) \%>\% print(n = Inf)

f <- function(x) stringr::str_remove(fs::path_file(x), ".final.log")
infs \%>\%
  lapply(ST_read_final_log, rename_col = f) \%>\%
  ST_merge_final_log()

}
