% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{ST_sj_tab}
\alias{ST_sj_tab}
\alias{ST_read_sj_tab}
\title{Read STAR SJ.out.tab file as a tibble}
\usage{
ST_read_sj_tab(fpath, parse = TRUE)
}
\arguments{
\item{fpath}{a path to the STAR final.log file}

\item{parse}{logical. If \code{TRUE} parse 3rd-5th columns. (default: \code{TRUE})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
See the section 5.5 "Splice junctions" in the STAR manual.
}
\examples{
infs <-
  system.file(package = "ngsmisc", "star") \%>\%
  fs::dir_ls(regexp = ".sj.tsv$")
ST_read_sj_tab(infs[1], parse = FALSE)
ST_read_sj_tab(infs[1], parse = TRUE)

# it can be converted to the `GenomicRanges::GRanges-class` object
# ST_read_sj_tab(infs[1], parse = TRUE) \%>\% plyranges::as_granges()

tbl_merge <-
  infs \%>\%
  lapply(ST_read_sj_tab) \%>\%
  lapply(head) \%>\%
  purrr::imap(~ dplyr::mutate(.x, sample = fs::path_file(.y))) \%>\%
  dplyr::bind_rows()
tbl_merge

# library(ggplot2)
# tbl_merge \%>\%
#   ggplot(aes(x = paste(seqnames, start, end, sep = "\n"), y = num_uniq_map_jc)) +
#   geom_col(aes(fill = sample, alpha = intron_motif)) +
#   theme_linedraw(base_size = 14) +
#   theme(panel.grid = element_blank()) +
#   labs(x = "Coordinates", y = "Number of uniquely mapped reads\ncrossing junction") +
#   scale_y_continuous(limits = c(0, NA), expand = expansion(c(0, .1))) +
#   scale_fill_viridis_d() +
#   scale_alpha_discrete(range = c(.5, 1))

}
