% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfiler.R
\name{wrap_strings}
\alias{wrap_strings}
\title{Wrap strings to defined width.}
\usage{
wrap_strings(string, w = 50)
}
\arguments{
\item{string}{A vector of strings.}

\item{w}{A number of string width to wrap.}
}
\description{
Wrap strings to defined width.
}
